feat(pages): owner-only Share button on served pages
Serving a private page to a viewer whose access capability belongs to the
page owner now injects a fixed-position Share anchor linking to the page's
card on the account dashboard. The pages origin still cannot call any API;
the button is navigation only, so the origin-isolation model is unchanged.
Share-link readers, invitees, and legacy pages get byte-identical output.

Dashboard cards gain id="page-<slug>" anchors with a :target highlight so
the deep link lands on the right card. Page ACL helpers move from
accounts.js into pages-acl.js to keep both under the export cap.

Closes #398
